Thread TypeScript conversion
